Added entities: Card, Product, Role Created repositories, services, and controllers for Card, Product, and Role entities Implemented global error handling mechanism Performed necessary debugging and error fixes

+++ back-end/src/main/java/com/techcareer/challenge/controller/UserController.java
@@ -2,10 +2,12 @@ package com.techcareer.challenge.controller;
 
 import com.techcareer.challenge.data.dto.UserDto;
 import com.techcareer.challenge.data.request.SignInRequest;
+import com.techcareer.challenge.exception.BadRequestException;
 import com.techcareer.challenge.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
 
 @RestController
 @RequestMapping("api/v1/user")
@@ -21,16 +23,25 @@ public class UserController {
 
     @PostMapping("/sign-in")
     public ResponseEntity<?> signIn(@RequestBody SignInRequest signInRequest){
+        if (signInRequest.getPassword().isEmpty() || signInRequest.getEmail().isEmpty()){
+            throw new BadRequestException("Invalid field");
+        }
         return ResponseEntity.ok(userService.signIn(signInRequest));
     }
 
     @PatchMapping("/update-user")
     public ResponseEntity<?> updateUser( @RequestParam Integer userId ,@RequestBody UserDto userDto){
+        if (userId < 0){
+            throw new BadRequestException("Invalid user ID");
+        }
         return ResponseEntity.ok(userService.updateUser(userId, userDto ));
     }
 
     @DeleteMapping("/delete-user")
     public void deleteUser(@RequestParam Integer userId){
+        if (userId < 0){
+            throw new BadRequestException("Invalid user ID");
+        }
        userService.deleteUser(userId);
     }
